Two query approach works correctly for board/pins
As mentioned in the previous commit message, there was a problem with
'one query for everything' approach while displaying user's rating for
a particular photo (if he has rated). So that problem as been resolved
by writing two queries for two scenarios and using the appropriate one
when needed. Now there are two things to work on..
1. Display of boardpins.jade
2. Making sure that if somone changes the URL while displaying board/pins
   then we get corret error messages (if required)