URL changed so that we can pass both userid and boardid while displaying board/boardpins
So far, the URL while displying a particular board's pins was passing only the boardid.
This design had a flaw in the case where we would wish to see other user's boards. So
now I pass both userid and boardid. This will be the design from now on. Whatever we do
we must pass the userid along with the URL. Also, the name change of 'board/pins' to
'board/boardpins' has also been done.
TODO
1. Create the photo page to display a particular photo in detail upon the click.