Code for Deleting pin complete.
Apparently there aren't any bugs but Delete Pin is not working. Delete
Pin page is getting all the parameters correctly and the query written
(has been tested in SQLDeveloper) also seems to be correct. But the pin
is not getting deleted. Also, Board pin count getting displayed is double
than the actual count.